Added a mapping to replace the word under the cursor
Writing a function that goes into insert mode, the user makes the change
and then everything else is changed looks hard (perhaps it's possible
with python instead of vimscript). So I made a mapping instead. The
problems are a) that it doesn't go into insert mode (can not see the
changes before making them) and b) does not replace the cursor position.