Added a better word replace function
Instead of creating a search command and waiting for the user to run it,
this function will let the user change the word he wants directly in the
text and then it will replace all other occurences as well. It looks
similar to the visual block change mode. It also restores the search
register and the cursor position.